Move pagination reducer into it's own folder
