+++ src/app/features/applications/application/log-stream-tab/log-stream-tab.component.spec.ts
@@ -5,7 +5,7 @@ import { EntityService } from '../../../../core/entity-service';
 import { ApplicationSchema, GetApplication } from '../../../../store/actions/application.actions';
 import { it } from '@angular/cli/lib/ast-tools/spec-utils';
 import { getInitialTestStoreState } from '../../../../test-framework/store-test-helper';
-import { paginationReducer } from '../../../../store/reducers/pagination.reducer';
+import { paginationReducer } from '../../../../store/reducers/pagination-reducer/pagination.reducer';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SharedModule } from '../../../../shared/shared.module';
 import { ApplicationEnvVarsService } from '../build-tab/application-env-vars.service';
