+++ src/app/shared/data-sources/list-data-source-cf.ts
@@ -10,7 +10,7 @@ import { Subscription } from 'rxjs/Subscription';
 import { schema } from 'normalizr';
 import { PaginationEntityState, PaginatedAction, QParam } from '../../store/types/pagination.types';
 import { AppState } from '../../store/app-state';
-import { getPaginationObservables, resultPerPageParam } from '../../store/reducers/pagination.reducer';
+import { getPaginationObservables, resultPerPageParam } from '../../store/reducers/pagination-reducer/pagination.reducer';
 import { AddParams, SetPage } from '../../store/actions/pagination.actions';
 import { ListDataSource } from './list-data-source';
 import { IListDataSource, getRowUniqueId } from './list-data-source-types';
