+++ src/app/store/reducers.module.ts
@@ -14,7 +14,7 @@ import { authReducer } from './reducers/auth.reducer';
 import { cnsisReducer } from './reducers/cnsis.reducer';
 import { createAppReducer } from './reducers/create-application.reducer';
 import { dashboardReducer } from './reducers/dashboard-reducer';
-import { paginationReducer } from './reducers/pagination.reducer';
+import { paginationReducer } from './reducers/pagination-reducer/pagination.reducer';
 import { uaaSetupReducer } from './reducers/uaa-setup.reducers';
 import { NgModule } from '@angular/core';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
