+++ src/app/store/reducers/pagination-reducer/pagination.reducer.ts
@@ -1,9 +1,9 @@
-import { PaginatedAction, PaginationEntityState, PaginationParam, QParam } from '../types/pagination.types';
+import { PaginatedAction, PaginationEntityState, PaginationParam, QParam } from '../../types/pagination.types';
 import { error } from 'util';
 import { Action, Store } from '@ngrx/store';
 import { denormalize, Schema } from 'normalizr';
 
-import { ApiActionTypes } from '../actions/request.actions';
+import { ApiActionTypes } from '../../actions/request.actions';
 import {
   ADD_PARAMS,
   AddParams,
@@ -15,13 +15,13 @@ import {
   SET_PARAMS,
   SetPage,
   SetParams,
-} from '../actions/pagination.actions';
-import { AppState } from '../app-state';
-import { mergeState } from './../helpers/reducer.helper';
+} from '../../actions/pagination.actions';
+import { AppState } from '../../app-state';
+import { mergeState } from '../../helpers/reducer.helper';
 import { Observable } from 'rxjs/Observable';
-import { selectPaginationState } from '../selectors/pagination.selectors';
-import { getRequestDataTypeState } from '../selectors/api.selectors';
-import { defaultCfEntitiesState } from '../types/entity.types';
+import { selectPaginationState } from '../../selectors/pagination.selectors';
+import { getRequestDataTypeState } from '../../selectors/api.selectors';
+import { defaultCfEntitiesState } from '../../types/entity.types';
 
 
 export const resultPerPageParam = 'results-per-page';
